Change ClientController into UserController
